pyproject.toml
[tool.poetry]
name = "chatter"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
pydantic = "^1.10.6"
fastapi = "^0.94.0"
python-multipart = "^0.0.6"
uvicorn = { extras = ["standard"], version = "^0.21.0" }
tortoise-orm = { extras = ["asyncpg", "accel"], version = "^0.19.3" }
aerich = "^0.7.1"
aiohttp = { extras = ["speedups"], version = "^3.8.1" }

openai-whisper = "^20230314"
numpy = "^1.23.0"
# remove at some point when llvmlite 0.34 isnt pulled anymore
llvmlite = "^0.39.1"
numba = "^0.56.4"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.0.2"
mypy = "^1.1.1"
black = "^23.1.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[tool.aerich]
tortoise_orm = "chatter.core.config.TORTOISE_ORM"
location = "./chatter/models/migrations"
src_folder = "./"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
